fix(procgroup): scope adopted reaping to Linux
