+++ muxcore/procgroup/process_tree_other.go
@@ -1,0 +1,7 @@
+//go:build !windows && !linux
+
+package procgroup
+
+func prepareProcessTreeAuthority() error { return nil }
+
+func drainWaitableGroupChildren(int) error { return nil }
