--- muxcore/supervisor/command_tree_subreaper_other_test.go
@@ -1,7 +1,0 @@
-//go:build !linux
-
-package supervisor
-
-import "testing"
-
-func prepareCommandTreeTest(*testing.T) {}
